Extract PreviewModeSelectorContainer
