+++ src/components/PreviewModeSelector.jsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import RadioGroup from './RadioGroup';
-import { setPreviewMode } from '../state/ui/actions';
 
 const PreviewModeSelector = (props) => {
   const { onOptionSelected, previewMode } = props;
@@ -22,13 +20,4 @@ PreviewModeSelector.propTypes = {
   onOptionSelected: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => {
-  const { previewMode } = state.ui;
-  return { previewMode };
-};
-
-const mapDispatchToProps = {
-  onOptionSelected: setPreviewMode,
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(PreviewModeSelector);
+export default PreviewModeSelector;
